Tpetra: Refactored specialized CrsMatrix (un)pack to use traits class.
This affects the specialized communication in Tpetra_Import_Util2.hpp,
NOT the usual pack and unpack routines in CrsMatrix.  The former now
uses a traits class to manage packing and unpacking.  This will help me
fix the Stokhos error that I introduced in previous commits (by getting
rid of the MatrixSerializationTraits class).