Tpetra: Refactored special CrsMatrix (un)pack to use Kokkos::View.
This affects the specialized communication in Tpetra_Import_Util2.hpp,
NOT the usual pack and unpack routines in CrsMatrix.  In my previous
commit, I made the former use a traits class (PackTraits) to manage
packing and unpacking.  In this commit, I changed PackTraits and the
code that uses it to use Kokkos::View (with MemoryUnmanaged) instead of
Teuchos::ArrayView.  This will help me fix the Stokhos error (see my
previous commit).