Cleaned up demangle() function usage in stk_util
There was some confusion about the demangle() function that STK
provides, where it was declared in:

  stk_util/diag/StringUtil.hpp

and defined in:

  stk_util/diag/Platform.cpp

as well as also being declared/defined in:

  stk_util/environment/Demangle.hpp
  stk_util/environment/Demangle.cpp

Only the version in environment was unit tested, and it was the version
in diag that was heavily used throughout the apps.  The duplicate
implementation in environment was removed in favor of the other, and the
implementation was moved to StringUtil.cpp, as expected by the end-users.

(From Sierra)