Trios: FullQueue test: remove assumption that Receive Queue is unbuffered
The FullQueue test assumed that the transports implemented the
Receive Queue without double buffering.  After the server creates a
Receive Queue with QUEUE_SIZE elements, the client expects the queue
to block after sending QUEUE_SIZE requests.  This is not the case
for the uGNI transport.  The FullQueue test no long makes this
assumption.  Instead, during phase 2 the client determines the
receive queue buffering capabilities and uses this information to
ensure the queue is blocked.