Kokkos: Added logic so that an unmanaged view can bind to tex objs
Added logic so that an unmanged view created from a raw pointer
can bind to texture memory if the memory was initally allocated by
Kokkos.  Improve error handling of cases where is is not possible to
bind a texture object.

Created a performance test to compare __ldg loads to text1Dfetch.
On kokkos-dev __ldg is faster with both (gcc 4.8.3 , cuda 6.5)
and (gcc 4.9.2 , cuda 7.0).  So I'm leaving in the code the forces
the __ldg command.