Kokkos: Added the possibility for compount types of execution/mem_space
This adds an Impl::DeviceInternal<ExecutionSpace,MemorySpace> class which
can be used to independently set execution_space and memory_space of a view.
This is internally done for the HostMirror of a view, so that the
HostMirror of a CudaUVMSpace view still uses the same memory space but has
the default execution space as its execution space.