Teuchos:  Fix type of "c" in BLAS::ROTG.
The Teuchos BLAS interface makes the incorrect assumption that
MagnitudeType == ScalarType if the ScalarType is not complex in its
declaration for the "c" argument to ROTG.  This doesn't work for Stokhos
embedded UQ scalar types.  I fixed this by deducing the type for "c"
from the internal GivensRotator class which is specialized on complex
versus real scalar types.  I also made isComplex be determined by
default for GivensRotator.  None of this should affect users with
standard scalar types, so it should be safe for backwards-compatibility.