Stokhos:  Setup MP::Vector/UQ::PCE unit tests to use MueLu better.
I created a utilities file for building a mean-based matrix and setup
the UQ::PCE unit tests to use this.  This better matches how we are
using the PCE scalar type.  I also changed the maximum coarse-grid size
to a smaller number so that more than one level is created in the
UQ::PCE unit tests.