Trios: MPI transport: fix RDMA tag conflict
The MPI transport assigns a unique set of tags to each registered
memory region to implement RDMA operations.  This commit refactors
the data_tag into a get_data_tag and a put_data_tag.  Without
separate tags, it is possible for GETS and PUTs to be intermingled
which leads to corrupted buffers.