Tpetra,Stokhos:  Provide overloads of getLocalDeepCopy.
I need to overload getLocalDeepCopy() which is used by
Tpetra::MultiVector for the UQ scalar types, for performance reasons on
Cuda.  To do that, I moved those functions and their associated functors
into a separate file, in the Tpetra::Details namespace.  This allows me
to overload them in Stokhos (in the future this should not be
necessary).  I also added pack/unpack kernels for the UQ::PCE.