Teuchos: Added __float128 {Serialization,Scalar,TypeName}Traits spec.
I added a __float128 specialization of SerializationTraits,
ScalarTraits, and TypeNameTraits to Teuchos.  This will help make it
possible to instantiate Tpetra classes with Scalar = __float128.

I also added a ScalarTraits<__float128> test.  This required a minor
change to the test.  It was using typeid(x) for x of various types.  I
found out that typeid(x) doesn't work if x has type __float128 (GCC
doesn't seem to implement the typeid operator for that type), so I
rewrote that part of the test to use Teuchos::TypeTraits::is_same<A,B>
instead.  That has the same effect, and doesn't change what the test
exercises.