Tpetra: Fixed build error when Kokkos refactor is disabled.
It's a minor build error in a test -- a missing include of
Kokkos_ArithTraits.hpp -- and it's only missing when the Kokkos
refactor version of Tpetra is disabled.  This commit adds the include
to the test.  It's harmless when the Kokkos refactor version of Tpetra
is enabled, as it is by default.