Tpetra::MultiVector: Removed deprecated createMultiVectorFromView.
The nonmember "constructor" function Tpetra::createMultiVectorFromView
took four arguments:

  - Teuchos::RCP<const Tpetra::Map> (the returned MultiVector's Map)
  - Teuchos::ArrayRCP<Scalar> (the data to view)
  - size_t LDA (Column stride)
  - size_t numVectors (number of columns)

and returned a Teuchos::RCP<Tpetra::MultiVector> which views (not
deep-copied) the data to which the Teuchos::ArrayRCP<Scalar> points.

This function has been deprecated for quite a while now.  It cannot
work with the new (Kokkos refactor) version of Tpetra, because Tpetra
now uses Kokkos::View and Kokkos::DualView for its data structures.
This means that Tpetra can't safely persistently view data to which an
ArrayRCP points.  The function also constrains the MultiVector's local
data to have column-major layout, which is something we want to relax
for various reasons.  Finally, the function is confusing and
redundant, because MultiVector already has constructors that view
input data (as Kokkos::DualView).  Thus, I have removed the deprecated
function.