Tpetra,Xpetra: Removed deprecated class Tpetra::ViewAccepter.
The Tpetra::ViewAccepter class was an implementation detail of
Tpetra::createMultiVectorFromView and Tpetra::createVectorFromView.
Both of those functions were deprecated a while back, and I recently
removed them, so there is no need for their implementation detail
class either.  Thus, I have removed Tpetra::ViewAccepter.

NOTE: One of Xpetra's code generation scripts referred to
Tpetra_ViewAccepter.hpp.  It was in an "ignore" section of the file,
which I'm guessing tells Xpetra not to process that file and generate
an Xpetra version of it.  However, I figured that if that header file
no longer exists, the script shouldn't refer to it, so I changed the
script accordingly.