TpetraCore: Only enable Lesson07 if Kokkos allows lambdas in dispatch.
The newly added Lesson 07 (thread-parallel fill into Kokkos data
structures, with hand-off to Tpetra data structures for a solve) uses
lambdas in parallel dispatch.  This feature is not yet available in
public releases of CUDA, even if C++11 is enabled.

This commit protects Lesson 07 with the CMake variable
Kokkos_ENABLE_CXX11_DISPATCH_LAMBDA, which I added in my previous
commit.