TpetraKernels: Starting to structure kernel package and use SEMI ETI
This starts to structure the Kernels package better. This is both good
for software engineering and will allow us to get better performance
as well.  The main reason for the performance improvement is to break
out BLAS kernels for common types and compile them in stand alone object
files.  This will give the optimizer more room to play as opposed to
compiling math kernels as part of one massive Tpetra object file (e.g.,
MV or CrsM).

"SEMI ETI" means that the technique does not actually depend on ETI.
It works whether or not ETI is enabled.  For each kernel, we will
provide both a generic implementation (exposed in a header file) and
full specializations.  Header files only see the declaration of the
full specializations; we hide their definitions in .cpp files.
Currently, we provide those specializations only for Scalar=double
and for all of Kokkos' device types.

The computational kernels have three "layers":

  1. Interface layer of functions in KokkosBlas namespace, that call

  2. Implementation layer of classes with static methods in
     KokkosBlas::Impl namespace, that invoke

  3. Functors in KokkosBlas::Impl namespace.

Users (and Tpetra) only interact with the outermost Layer 1.
Layer 1 consolidates template arguments into those that Layer 2
would want to use for good performance.  This also reduces the
number of explicit instantiations of the actual kernels.  For
example, if users pass a View<double*> into Layer 1, it turns
that into a View<const double*, ...> and hands it off to Layer 2.

Each class in Layer 2 has a generic version.  It may also have a full
(or even a partial) specialization.  Full specializations live in their
own .cpp files.  Each kernel (e.g., dot product) gets its own .cpp file.
This keeps the files short so that the compiler can optimize.

The specializations in Layer 2 still use the same generic functors from
Layer 3.  In particular, we do NOT write a custom kernel for each Scalar
type.  It is still possible to do that, but we prefer not to do that
unless we have evidence that it is required for performance.  Our
experiments show that it is not.

Here is how we name the new files:

  - Kokkos_BLAS{1,2,3}.hpp contain Layer 1, for general vectors or
    matrices
  - Kokkos_BLAS{1,2,3}_MV.hpp contain variants of Layer 1 kernels
    appropriate for multivectors (as in the 2-D Views used by
    Tpetra::MultiVector)
  - impl/Kokkos_BLAS{1,2,3}[_MV]_impl.hpp contains Layer 2 and
    generic Layer 3 for Kokkos_BLAS{1,2,3}[_MV].hpp
  - impl/Kokkos_BLAS{1,2,3}[_MV]_impl_$KERNEL.cpp contains Layer 2
    full specialization definitions (instantiations), as declared
    in impl/Kokkos_BLAS{1,2,3}[_MV]_impl.hpp