TpetraKernels: Added 3-arg nrm2_squared() overload for X(:,X_col).
Tpetra::MultiVector::norm2() would like to be able to select columns
of the input multivectors on which to compute a 2-norm.  This makes
norm2() work correctly with so-called "noncontiguous views" (i.e.,
Tpetra views of multivector columns that are not contiguous).

This commit adds a 3-argument overload of KokkosBlas::nrm2_squared()
that lets callers select the column of X (X_col) to use.  It still
accepts Kokkos Views of the same ranks as the 2-argument version of
nrm2_squared() -- rank 1 for the output array, and rank 2 for the
input multivector X.  This may help make the implementation of
Tpetra::MultiVector::norm2 shorter.