TpetraKernels: Re-enabled sparse matrix-vector multiply test.
The sparse mat-vec test used to depend on Gtest.  I was getting link
errors related to Gtest and Pthreads (see my previous commit message).
The test really didn't _need_ Gtest, and TpetraKernels already depends
on TeuchosCore.  Thus I changed the test to depend on the Teuchos unit
test framework instead, and re-enabled the test.

If people prefer that it be stand-alone, it would be easy to do that
at some later point.  For now, it's more important that the test
actually builds.