Kokkos/Tpetra: Fixed performance bug for threaded execution of CrsMatVec
When pulling out the old vectorization interface I screwed up the
calculation of the number of teams and number of threads per team.
That caused the previous issue with overflows of shorts. It also
caused a severe performance regression in threaded mode.

This commit fixes it. To facilitate the fix I added a
team_size_recommended(Functor,VectorLength) function to
the TeamPolicy. This is to get the recommended team size
for a 3D executin policy with a vector length. This currently
just divides the recommended team_size by vector lenght when
using Cuda.