TeuchosKokkosCompat: Fixed residual .cu use in CMakeLists.txt.
The KokkosCompat subpackage of Teuchos was attempting to build a
teuchoskokkoscompat_cuda library if Kokkos_ENABLE_Cuda was ON.
We don't need to deal with .cu files separately anymore, due to
the nvcc_wrapper script.  Furthermore, this broke the CUDA build
for me on Shannon.  When I took out that code from the
CMakeLists.txt file, the build worked for me.